Updating project structure
